requirements.txt
streamlit
google-generativeai
python-dotenv
requests
langdetect
SpeechRecognition
gTTS
pydub
pyaudio
